+++ components/Tags.lua
@@ -47,20 +47,9 @@ function tags:init()
 	self.sortedTags = {}
 	self:setSortedTags()
 
-	local mountOptionsMenu = MountJournal.mountOptionsMenu
-	util.setMixin(mountOptionsMenu, MJDropDownButtonMixin)
-	mountOptionsMenu:ddSetInit(function(...) self:mountOptionsMenu_Init(...) end, "menu")
-
-	MountJournal_ShowMountDropdown = function(index, anchorTo, offsetX, offsetY)
-		if not index then return end
-		local _,_,_, active, isUsable, _,_,_,_,_,_, mountID = C_MountJournal.GetDisplayedMountInfo(index)
-		MountJournal.menuMountIndex = index
-		MountJournal.menuMountID = mountID
-		MountJournal.active = active
-		MountJournal.menuIsUsable = isUsable
-		mountOptionsMenu:dropDownToggle(1, nil, anchorTo, offsetX, offsetY)
-		PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
-	end
+	self.mountOptionsMenu = MountJournal.mountOptionsMenu
+	util.setMixin(self.mountOptionsMenu, MJDropDownButtonMixin)
+	self.mountOptionsMenu:ddSetInit(function(...) self:mountOptionsMenu_Init(...) end, "menu")
 end
 
 
@@ -91,13 +80,76 @@ function tags:resetFilter()
 end
 
 
+function tags:dragButtonClick(btn, mouseBtn)
+	local parent = btn:GetParent()
+	if mouseBtn ~= "LeftButton" then
+		self:showMountDropdown(parent.index, btn, 0, 0)
+	elseif IsModifiedClick("CHATLINK") then
+		local id = parent.spellID
+		if MacroFrame and MacroFrame:IsShown() then
+			local spellName = GetSpellInfo(id)
+			ChatEdit_InsertLink(spellName)
+		else
+			local spellLink = GetSpellLink(id)
+			ChatEdit_InsertLink(spellLink)
+		end
+	else
+		C_MountJournal.Pickup(parent.index)
+	end
+end
+
+
+do
+	local lastMountClick = 0
+	function tags:listItemClick(btn, mouseBtn)
+		if mouseBtn ~= "LeftButton" then
+			self:showMountDropdown(btn.index, btn, 0, 0)
+		elseif IsModifiedClick("CHATLINK") then
+			local id = btn.spellID
+			if MacroFrame and MacroFrame:IsShown() then
+				local spellName = GetSpellInfo(id)
+				ChatEdit_InsertLink(spellName)
+			else
+				local spellLink = GetSpellLink(id)
+				ChatEdit_InsertLink(spellLink)
+			end
+		else
+			local time = GetTime()
+			if btn.mountID ~= MountJournal.selectedMountID then
+				MountJournal_SetSelected(btn.mountID, btn.spellID, btn)
+			elseif time - lastMountClick < .4 then
+				local _,_,_, active, isUsable, _,_,_,_,_,_, mountID = C_MountJournal.GetDisplayedMountInfo(btn.index)
+				if active then
+					C_MountJournal.Dismiss()
+				elseif isUsable then
+					C_MountJournal.SummonByID(mountID)
+				end
+			end
+			lastMountClick = time
+		end
+	end
+end
+
+
+function tags:showMountDropdown(index, anchorTo, offsetX, offsetY)
+	if not index then return end
+	local _,_,_, active, isUsable, _,_,_,_,_,_, mountID = C_MountJournal.GetDisplayedMountInfo(index)
+	MountJournal.menuMountIndex = index
+	MountJournal.menuMountID = mountID
+	MountJournal.active = active
+	MountJournal.menuIsUsable = isUsable
+	self.mountOptionsMenu:dropDownToggle(1, nil, anchorTo, offsetX, offsetY)
+	PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
+end
+
+
 function tags:mountOptionsMenu_Init(btn, level)
 	if not MountJournal.menuMountIndex then return end
 	local info = {}
 	local mountIndex, mountID = MountJournal.menuMountIndex, MountJournal.menuMountID
 
 	if level == 1 then
-		local _,_,_, active, _,_, isFavorite = C_MountJournal.GetMountInfoByID(mountID)
+		local _,_,_, active, _,_, isFavorite, _,_,_, isCollected = C_MountJournal.GetMountInfoByID(mountID)
 		local needsFanfare = C_MountJournal.NeedsFanfare(mountID)
 		info.notCheckable = true
 
@@ -121,7 +173,7 @@ function tags:mountOptionsMenu_Init(btn, level)
 
 		if not needsFanfare then
 			local _, canFavorite = C_MountJournal.GetIsFavorite(mountIndex)
-			info.disabled = not canFavorite
+			info.disabled = not (isCollected and canFavorite)
 
 			if isFavorite then
 				info.text = BATTLE_PET_UNFAVORITE